refactor: Repository 인터페이스를 domain 패키지로 이동
