+++ src/main/java/kitchenpos/menu/application/request/MenuRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.menu.application.request;
 
-import kitchenpos.menuproduct.application.request.MenuProductRequest;
+import kitchenpos.menuproduct.request.MenuProductRequest;
 
 import java.math.BigDecimal;
 import java.util.List;
