+++ src/main/java/kitchenpos/order/OrderLineItemRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.order.application;
+package kitchenpos.order;
 
-import kitchenpos.order.OrderLineItem;
 import org.springframework.data.repository.Repository;
 
 import java.util.List;
