+++ src/main/java/kitchenpos/ordertable/OrderTableValidator.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.application;
+package kitchenpos.ordertable;
 
 import java.util.List;
