+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -3,14 +3,15 @@ package kitchenpos.menu.application;
 import kitchenpos.menu.Menu;
 import kitchenpos.menu.MenuName;
 import kitchenpos.menu.MenuPrice;
+import kitchenpos.menu.MenuRepository;
 import kitchenpos.menu.application.request.MenuRequest;
-import kitchenpos.menugroup.application.MenuGroupRepository;
-import kitchenpos.menuproduct.application.MenuProductRepository;
-import kitchenpos.menuproduct.application.request.MenuProductRequest;
+import kitchenpos.menugroup.MenuGroupRepository;
+import kitchenpos.menuproduct.MenuProductRepository;
+import kitchenpos.menuproduct.request.MenuProductRequest;
 import kitchenpos.product.Product;
 import kitchenpos.product.ProductName;
 import kitchenpos.product.ProductPrice;
-import kitchenpos.product.application.ProductRepository;
+import kitchenpos.product.ProductRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
