+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -3,10 +3,12 @@ package kitchenpos.order.application;
 import kitchenpos.menu.Menu;
 import kitchenpos.menu.MenuName;
 import kitchenpos.menu.MenuPrice;
-import kitchenpos.menu.application.MenuRepository;
+import kitchenpos.menu.MenuRepository;
 import kitchenpos.order.Order;
 import kitchenpos.order.OrderLineItem;
 import kitchenpos.order.OrderLineItemQuantity;
+import kitchenpos.order.OrderLineItemRepository;
+import kitchenpos.order.OrderRepository;
 import kitchenpos.order.OrderStatus;
 import kitchenpos.order.application.request.OrderLineItemDto;
 import kitchenpos.order.application.request.OrderRequest;
@@ -14,7 +16,7 @@ import kitchenpos.order.application.request.OrderStatusRequest;
 import kitchenpos.ordertable.Empty;
 import kitchenpos.ordertable.NumberOfGuests;
 import kitchenpos.ordertable.OrderTable;
-import kitchenpos.ordertable.application.OrderTableRepository;
+import kitchenpos.ordertable.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
