+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -3,6 +3,7 @@ package kitchenpos.product.application;
 import kitchenpos.product.Product;
 import kitchenpos.product.ProductName;
 import kitchenpos.product.ProductPrice;
+import kitchenpos.product.ProductRepository;
 import kitchenpos.product.application.request.ProductRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
